pyproject.toml
[project]
name = "boostcamp-ai-tech-product-serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow[celery]==3.1.5",
]
